chore: move impl into exec/scan
