+++ polars/polars-lazy/src/physical_plan/executors/scan/mod.rs
@@ -2,6 +2,8 @@
 mod csv;
 #[cfg(feature = "ipc")]
 mod ipc;
+#[cfg(feature = "json")]
+mod json;
 #[cfg(feature = "parquet")]
 mod parquet;
